Create & display appointments [Issue 16, Issue 18] (#37)
* Create AppointmentBox

Displays an overview of the session details:
- the appointment name
- the therapist
- the duration of the session.

It's position and size depends on the duration, but position changing based on start time hasn't been implemented yet.

* Appointments adjust position on the calendar based on it's start time

- AppointmentBox takes in the height of each row and the start hour in the calendar to calculate its position

* Calendar displays appointments in a set week

- Calendar takes a list of sessions & a day of the week that is to be displayed. It sorts the given appointments into the days
- If a day is outside of the week then it gets ignored

* Fixed AppointmentBox position

fixed an issue where if there were multiple sessions in the same day, the position for some of the boxes would be wrong

* Change calendar hours based on appointments

Calendar calculates the earliest start hour and the end hour from all the sessions that are to be displayed and displays the time accordingly

* Update WeekViewCalendar.vue

* Updated appointmentbox & calendar

As far as sessions with shorter durations, I made it so it would cut out the name and time info, but would it be better to have a minimum height instead?

* Appointment scales text size with box size

When the session < 50, the format changes to have the name on the left and time on the right. Therapist name gets cut out at < 50 to have space for the session name in case it's long

* Center text in the appointment box